feat: improve folder structure and extract components
